+++ lib/core/constants/strings.dart
@@ -1,0 +1,6 @@
+class Strings {
+  // App
+  static const appTitle = 'Kanban';
+  static const homePage = 'Kanban Home Page';
+  static const appVersion = '1.0.1';
+}