+++ CHANGELOG.md
@@ -25,6 +25,21 @@ Here we write the updates of future releases.
 
 
 
+## [1.0.1] - 12-05-2024
+
+### Added
+- Constants folder to maintain all the constant values like strings images and booleans
+- Themes folder to maintain the themes
+- Cubits folder to store all the general cubits
+- Theme cubit to maintain the theme of the app
+- Router folder to store all the routes
+- Utils folder to maintain all the required utilities like app texts and custom print function
+
+### Changed
+- Initial routing to splash screen using named routes
+
+
+
 ## [1.0.0] - 12-05-2024
 
 ### Added